Split system purpose tests to a separate file
The main test file is getting a bit too big, let's split
system purpose helper module tests to a separate file.

(cherry picked from commit 43a4e4123d7e5cc3d130e869de2ed2d6bc9bc267)

Related: rhbz#1951709
